updated cli version and moved helper scripts out of src/ditto
