+++ package.json
@@ -8,7 +8,7 @@
     "build": "tsc && vite build",
     "lint": "eslint . --ext ts,tsx --report-unused-disable-directives --max-warnings 0",
     "preview": "vite preview",
-    "ditto:pull": "cd src && ditto-cli pull && npx tsx ditto/fetchVariables.ts"
+    "ditto:pull": "npx tsx ditto-helpers/fetchVariables.ts && cd src && ditto-cli pull --sample-data"
   },
   "dependencies": {
     "@types/js-yaml": "^4.0.9",
@@ -19,7 +19,7 @@
     "wouter": "^2.12.0"
   },
   "devDependencies": {
-    "@dittowords/cli": "^4.1.0",
+    "@dittowords/cli": "^4.4.0",
     "@types/react": "^18.2.15",
     "@types/react-dom": "^18.2.7",
     "@typescript-eslint/eslint-plugin": "^6.0.0",
