+++ tsconfig.json
@@ -20,6 +20,6 @@
     "noUnusedParameters": true,
     "noFallthroughCasesInSwitch": true
   },
-  "include": ["src"],
+  "include": ["src", "ditto-helpers/helpers.ts", "ditto-helpers/fetchVariables.ts"],
   "references": [{ "path": "./tsconfig.node.json" }]
 }
